apps/backend/lifeos-rag-api/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
attrs==25.4.0
banks==2.2.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
ecdsa==0.19.1
fastapi==0.124.2
fastapi-limiter==0.1.6
filetype==1.2.0
frozenlist==1.8.0
fsspec==2025.12.0
greenlet==3.3.0
griffe==1.15.0
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.0
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
llama-cloud==0.1.35
llama-cloud-services==0.6.54
llama-index==0.14.10
llama-index-cli==0.5.3
llama-index-core==0.14.10
llama-index-embeddings-openai==0.5.1
llama-index-graph-stores-neo4j==0.5.1
llama-index-indices-managed-llama-cloud==0.9.4
llama-index-instrumentation==0.4.2
llama-index-llms-openai==0.6.10
llama-index-readers-file==0.5.5
llama-index-readers-llama-parse==0.5.1
llama-index-vector-stores-qdrant==0.9.0
llama-index-workflows==2.11.5
llama-parse==0.6.54
MarkupSafe==3.0.3
marshmallow==3.26.1
multidict==6.7.0
mypy_extensions==1.1.0
neo4j==5.28.2
nest-asyncio==1.6.0
networkx==3.6.1
nltk==3.9.2
numpy==2.3.5
openai==2.11.0
packaging==25.0
pandas==2.2.3
passlib==1.7.4
pillow==12.0.0
platformdirs==4.5.1
portalocker==3.2.0
propcache==0.4.1
protobuf==6.33.2
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
pypdf==6.4.1
python-arango==8.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
pytz==2025.2
PyYAML==6.0.3
qdrant-client==1.16.1
redis==7.1.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.45
starlette==0.50.0
striprtf==0.0.26
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.2
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
wrapt==1.17.3
yarl==1.22.0
zipp==3.23.0